requirements.txt
# === Core ===
python-dotenv
loguru

# === Voice Interface ===
vosk
pyttsx3
pvporcupine

# === ML Classifier ===
scikit-learn
nltk
numpy

# === Agentic AI + LLM ===
langchain
langgraph
ollama
langchain-community

# === Vector Store for RAG ===
chromadb
sentence-transformers
tqdm

# === System Integration ===
pyautogui
psutil
# === Miscellaneous ===
requests
# === Testing ===
pytest
